Consolidate legacy type analysis pipeline
